requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
faiss-cpu
sentence-transformers
transformers
openai
pdfminer.six
python-multipart
gunicorn
nltk
pymilvus
PyMuPDF
requests
beautifulsoup4
readability-lxml
html2text
tiktoken
spacy